Moving canvas+svg under render
