+++ src/render/canvas/index.js
