+++ src/render/canvas/path.js
@@ -1,6 +1,6 @@
 define(function(require, module, exports) {
   var d3 = require('d3'),
-      Bounds = require('../core/Bounds');
+      Bounds = require('../../core/Bounds');
 
   // Path parsing and rendering code taken from fabric.js -- Thanks!
   var cmdLength = { m:2, l:2, h:1, v:1, c:6, s:4, q:4, t:2, a:7 },
