+++ src/render/svg/index.js
